pyproject.toml
[tool.poetry]
name = "swarm-stacks"
version = "2020.5.19"
description = "Builders for common Docker swarm stacks"
authors = ["IaroslavR <iarruss@ya.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
invoke = "^1.4.1"
loguru = "^0.4.1"
pyyaml = "^5.3.1"
jinja2 = "^2.11.2"
# dev deps
black = { version = "^19.10b0", optional = true }
coverage = {version = "^5.1", optional = true }
darglint = { version = "^1.2.3", optional = true }
flake8 = { version = "^3.7.9", optional = true }
flake8-annotations = { version = "^2.1.0", optional = true }
flake8-bandit = { version = "^2.1.2", optional = true }
flake8-black = { version = "^0.1.1", optional = true }
flake8-bugbear = { version = "^20.1.4", optional = true }
flake8-docstrings = { version = "^1.5.0", optional = true }
flake8-import-order = { version = "^0.18.1", optional = true }
pytest = { version = "^5.4.1", optional = true }
pytest-flake8 = { version = "^1.0.5", optional = true }
safety = { version = "^1.8.7", optional = true }

[tool.poetry.extras]
dev = [
    "black", "coverage", "darglint", "flake8", "flake8-annotations",
    "flake8-bandit", "flake8-black", "flake8-bugbear", "flake8-docstrings",
    "flake8-import-order", "pytest", "safety", "pytest-flake8"]

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
portainer-stack = "swarm_stacks.portainer:program.run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
